+++ plugins/woocommerce-admin/packages/admin-e2e-tests/src/specs/activate-and-setup/complete-onboarding-wizard.ts
@@ -6,6 +6,8 @@ import { WcHomescreen } from '../../pages/WcHomescreen';
 import { TaskTitles } from '../../constants/taskTitles';
 import { Login } from '../../pages/Login';
 import { WcSettings } from '../../pages/WcSettings';
+import { ProductsSetup } from '../../pages/ProductsSetup';
+import { deactivateAndDeleteExtension } from '../../utils/actions';
 
 /* eslint-disable @typescript-eslint/no-var-requires */
 const {
@@ -349,8 +351,175 @@ const testDifferentStoreCurrenciesWCPay = () => {
 	} );
 };
 
+const testSubscriptionsInclusion = () => {
+	describe( 'A non-US store will not see the Subscriptions inclusion', () => {
+		const profileWizard = new OnboardingWizard( page );
+		const login = new Login( page );
+
+		beforeAll( async () => {
+			await login.login();
+		} );
+
+		it( 'can complete the store details section', async () => {
+			await profileWizard.navigate();
+			await profileWizard.storeDetails.completeStoreDetailsSection( {
+				countryRegionSubstring: 'fran',
+				countryRegionSelector: 'FR',
+				countryRegion: 'France',
+			} );
+
+			// Wait for "Continue" button to become active
+			await profileWizard.continue();
+
+			// Wait for usage tracking pop-up window to appear
+			await profileWizard.optionallySelectUsageTracking();
+		} );
+
+		it( 'can complete the product types section, Subscriptions copy is not visible', async () => {
+			// Query for the industries checkboxes
+			await profileWizard.industry.isDisplayed();
+			await profileWizard.industry.uncheckIndustries();
+			await profileWizard.industry.selectIndustry( 'Health and beauty' );
+			await profileWizard.continue();
+			await profileWizard.productTypes.isDisplayed( 7 );
+			await profileWizard.productTypes.uncheckProducts();
+			await profileWizard.productTypes.selectProduct( 'Subscriptions' );
+			await expect( page ).not.toMatchElement( 'p', {
+				text:
+					'The following extensions will be added to your site for free: WooCommerce Payments. An account is required to use this feature.',
+			} );
+
+			await profileWizard.continue();
+			await page.waitForNavigation( { waitUntil: 'networkidle0' } );
+		} );
+
+		it( 'can complete the business details tab', async () => {
+			await profileWizard.business.isDisplayed();
+
+			await profileWizard.business.selectProductNumber(
+				config.get( 'onboardingwizard.numberofproducts' )
+			);
+			await profileWizard.business.selectCurrentlySelling(
+				config.get( 'onboardingwizard.sellingelsewhere' )
+			);
+
+			await profileWizard.continue();
+		} );
+
+		it( 'should display the WooCommerce Payments extension after it has been installed', async () => {
+			await profileWizard.business.freeFeaturesIsDisplayed();
+			await profileWizard.business.expandRecommendedBusinessFeatures();
+
+			expect( page ).toMatchElement( 'a', {
+				text: 'WooCommerce Payments',
+			} );
+		} );
+
+		it( 'should display the task "Add Subscriptions to my store"', async () => {
+			await profileWizard.goToOBWStep( 'Store Details' );
+			await profileWizard.skipStoreSetup();
+			const homescreen = new WcHomescreen( page );
+			await homescreen.isDisplayed();
+			await homescreen.possiblyDismissWelcomeModal();
+			const tasks = await homescreen.getTaskList();
+			expect( tasks ).toContain( 'Add Subscriptions to my store' );
+		} );
+
+		it( 'can select the Subscription option in the "Start with a template" modal', async () => {
+			const productsSetup = new ProductsSetup( page );
+			await productsSetup.navigate();
+			await productsSetup.isDisplayed();
+			await productsSetup.clickStartWithTemplate();
+			await productsSetup.isStartWithATemplateDisplayed( 3 );
+		} );
+	} );
+	describe( 'A US store will see the Subscriptions inclusion', () => {
+		const profileWizard = new OnboardingWizard( page );
+		const login = new Login( page );
+
+		afterAll( async () => {
+			await deactivateAndDeleteExtension( 'woocommerce-payments' );
+			await login.logout();
+		} );
+
+		it( 'can complete the store details section', async () => {
+			await profileWizard.navigate();
+			await profileWizard.storeDetails.completeStoreDetailsSection( {
+				countryRegionSubstring: 'cali',
+				countryRegionSelector: 'US\\:CA',
+				countryRegion: 'United States (US) — California',
+			} );
+
+			// Wait for "Continue" button to become active
+			await profileWizard.continue();
+
+			// Wait for usage tracking pop-up window to appear
+			await profileWizard.optionallySelectUsageTracking();
+		} );
+
+		it( 'can complete the product types section, the Subscriptions copy now is visible', async () => {
+			// Query for the industries checkboxes
+			await profileWizard.industry.isDisplayed();
+			await profileWizard.industry.uncheckIndustries();
+			await profileWizard.industry.selectIndustry( 'Health and beauty' );
+			await profileWizard.continue();
+			await profileWizard.productTypes.isDisplayed( 7 );
+			await profileWizard.productTypes.uncheckProducts();
+			await profileWizard.productTypes.selectProduct( 'Subscriptions' );
+			await expect( page ).toMatchElement( 'p', {
+				text:
+					'The following extensions will be added to your site for free: WooCommerce Payments. An account is required to use this feature.',
+			} );
+
+			await profileWizard.continue();
+			await page.waitForNavigation( { waitUntil: 'networkidle0' } );
+		} );
+
+		it( 'can complete the business details tab', async () => {
+			await profileWizard.business.isDisplayed();
+
+			await profileWizard.business.selectProductNumber(
+				config.get( 'onboardingwizard.numberofproducts' )
+			);
+			await profileWizard.business.selectCurrentlySelling(
+				config.get( 'onboardingwizard.sellingelsewhere' )
+			);
+
+			await profileWizard.continue();
+		} );
+
+		it( 'cannot see the WooCommerce Payments extension after it has been installed', async () => {
+			await profileWizard.business.freeFeaturesIsDisplayed();
+			await profileWizard.business.expandRecommendedBusinessFeatures();
+
+			expect( page ).not.toMatchElement( 'a', {
+				text: 'WooCommerce Payments',
+			} );
+		} );
+
+		it( 'should not display the task "Add Subscriptions to my store"', async () => {
+			await profileWizard.goToOBWStep( 'Store Details' );
+			await profileWizard.skipStoreSetup();
+			const homescreen = new WcHomescreen( page );
+			await homescreen.isDisplayed();
+			await homescreen.possiblyDismissWelcomeModal();
+			const tasks = await homescreen.getTaskList();
+			expect( tasks ).not.toContain( 'Add Subscriptions to my store' );
+		} );
+
+		it( 'can select the Subscription option in the "Start with a template" modal', async () => {
+			const productsSetup = new ProductsSetup( page );
+			await productsSetup.navigate();
+			await productsSetup.isDisplayed();
+			await productsSetup.clickStartWithTemplate();
+			await productsSetup.isStartWithATemplateDisplayed( 4 );
+		} );
+	} );
+};
+
 module.exports = {
 	testAdminOnboardingWizard,
 	testSelectiveBundleWCPay,
 	testDifferentStoreCurrenciesWCPay,
+	testSubscriptionsInclusion,
 };