+++ plugins/woocommerce-admin/tests/e2e/specs/activate-and-setup/complete-onboarding-wizard.test.tsx
@@ -2,8 +2,10 @@ const {
 	testAdminOnboardingWizard,
 	testSelectiveBundleWCPay,
 	testDifferentStoreCurrenciesWCPay,
+	testSubscriptionsInclusion,
 } = require( '@woocommerce/admin-e2e-tests' );
 
 testAdminOnboardingWizard();
 testSelectiveBundleWCPay();
 testDifferentStoreCurrenciesWCPay();
+testSubscriptionsInclusion();